requirements.txt
# Backend principal
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.1

# Procesamiento de imágenes y videos
Pillow==10.0.0
opencv-python-headless==4.8.0.76
moviepy==1.0.3
imageio-ffmpeg==0.4.8

# Obtención de noticias
newspaper3k==0.2.8
google-news==1.6.9
newsapi-python==0.2.7
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Procesamiento de texto
sumy==0.10.0

# Caché y rendimiento
redis==4.5.5
python-memcached==1.59

# Utilidades
tqdm==4.66.1  # Para barras de progreso
python-slugify==8.0.1

# Seguridad
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# Pruebas
pytest==7.4.2
pytest-asyncio==0.21.1
httpx==0.24.1

# Documentación
mkdocs==1.5.2
mkdocs-material==9.2.5b1
